Build prometheus.Histogram instead of Observer
The unprecise usage of `prometheus.Observer` instead of
`prometheus.Histogram` may be misleading, since both `Histogram` and
`Summary` interfaces embed the `Observer` interface.

This is a breaking change since previous code won't compile with this
version, so there's no better moment (except for v2) for doing this than
going v1.