Histograms: require buckets to be explicit
Default buckets were removed and buckets are now required: writing the
buckets is a 5 second effort while forgetting about them and having
wrong buckets in production is a lot of work (and maybe even an incident
if wrong buckets don't provide enough visibility on an issue)

So now any code without buckets on the histograms will fail to intialize
that metrics.

Finally, empty buckets tag is now allowed and it builds an empty buckets
slice, which will cause prometheus library to use its own default
buckets.

Notice that this change identified an issue in one of our own tests
where buckets tag was wrong (`metricsbuckets`) and it was unnoticed
because default buckets where used instead.